task1_media_drf-main/requirements.txt
asgiref==3.5.2
Django==4.1.3
djangorestframework==3.13.1
eureka==0.15
ifaddr==0.2.0
pymongo==3.12.2
pytz==2022.1
sqlparse==0.2.4
uuid==1.30
psycopg2-binary
py_eureka_client
django-environ